Add files via upload
Food insecurity is a critical issue that affects millions of people worldwide, with significant implications for health, economic stability, and social welfare. In South Africa, food insecurity is particularly prevalent, exacerbated by various factors such as economic inequality, unemployment, and environmental challenges. To address this issue, a comprehensive analysis was conducted, focusing on the nine provinces of South Africa over the years 2021, 2022, and 2023.   Power BI, a powerful business analytics tool, was employed to manage and visualize the extensive data collected. The tool was used for data extraction, transformation, loading, and publishing, allowing for a thorough and efficient analysis of the complex data involved. The visualizations created in Power BI provided clear insights into the trends and factors contributing to food 
insecurity across different regions, enabling the identification of key areas for intervention.  A significant aspect of the project was the development and implementation of strategies to mitigate food insecurity. Another plan of action involved the calculation of Key Performance Indicators (KPIs) to measure the total food supply and assess the risk of food insecurity. This quantitative approach provided a clear framework for monitoring progress and making data-driven decisions.